Create view pager to hold home frag and list frag
Replace SyncFrag with SyncAct
