+++ app/src/main/java/com/detroitlabs/devicemanager/ui/HomeFragment.java
@@ -6,7 +6,6 @@ import android.arch.lifecycle.ViewModelProviders;
 import android.os.Build;
 import android.os.Bundle;
 import android.support.annotation.Nullable;
-import android.support.v4.app.FragmentTransaction;
 import android.transition.TransitionManager;
 import android.view.KeyEvent;
 import android.view.LayoutInflater;
@@ -16,7 +15,6 @@ import android.view.ViewGroup;
 import android.view.inputmethod.EditorInfo;
 import android.widget.TextView;
 
-import com.detroitlabs.devicemanager.R;
 import com.detroitlabs.devicemanager.databinding.FragHomeBinding;
 import com.detroitlabs.devicemanager.db.Device;
 import com.detroitlabs.devicemanager.utils.ViewUtil;
@@ -100,12 +98,7 @@ public class HomeFragment extends LifecycleFragment {
         binding.buttonOtherDevices.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                DeviceListFragment deviceListFragment = DeviceListFragment.newInstance();
-                FragmentTransaction ft = getFragmentManager().beginTransaction();
-                ft.add(R.id.container, deviceListFragment);
-                ft.setTransition(FragmentTransaction.TRANSIT_FRAGMENT_CLOSE);
-                ft.addToBackStack(null);
-                ft.commit();
+                // TODO: 7/24/17 swipe to list page
             }
         });
     }
