+++ app/src/main/res/layout/activity_sync.xml
@@ -1,0 +1,25 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools">
+
+    <RelativeLayout
+        android:layout_width="match_parent"
+        android:layout_height="match_parent">
+
+        <TextView
+            android:id="@+id/status_text"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_centerInParent="true"
+            tools:text="Sync database" />
+
+        <ProgressBar
+            android:id="@+id/progressBar"
+            style="@style/HorizontalProgressBar"
+            android:layout_width="100dp"
+            android:layout_height="wrap_content"
+            android:layout_below="@id/status_text"
+            android:layout_centerHorizontal="true"
+            android:indeterminate="true" />
+    </RelativeLayout>
+</layout>
