+++ app/src/main/res/layout/frag_home.xml
@@ -76,6 +76,8 @@
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:padding="@dimen/normal_padding"
+                android:descendantFocusability="beforeDescendants"
+                android:focusableInTouchMode="true"
                 tools:visibility="visible">
 
                 <Button
